% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseflow_sep.R
\name{baseflow_sep}
\alias{baseflow_sep}
\title{Baseflow separation}
\usage{
baseflow_sep(x, method = "LH", min_segment_length = 30, ...)
}
\arguments{
\item{x}{A tibble.}

\item{method}{A string specifying the method to use.}

\item{min_segment_length}{An integer specifying the
minimum number of consecutive time points for which to
perform baseflow separation.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A tibble
}
\description{
Baseflow separation
}
\examples{
\dontrun{
print("Hello, world")
}
}
