% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kge.R
\name{kge2012}
\alias{kge2012}
\alias{kge2012.data.frame}
\alias{kge2012_vec}
\title{Modified Kling-Gupta Efficiency (KGE')}
\usage{
kge2012(data, ...)

\method{kge2012}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)

kge2012_vec(truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
and \code{estimate} arguments.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{kge2012_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
Calculate the modified Kling-Gupta Efficiency (\emph{Kling et al., 2012}),
aka \eqn{KGE'}. Dimensionless (from \eqn{-\infty} to 1).
\code{kge2012()} assesses the accuracy of
simulated data by considering correlation, bias, and variability relative
to observed data.
}
\details{
The Modified Kling-Gupta Efficiency is a composite metric that decomposes
model performance into three components: correlation (\eqn{r}),
bias ratio (\eqn{\beta}), and variability ratio (\eqn{\gamma}).
It improves upon the Kling-Gupta Efficiency (see \link{kge}) by replacing
standard deviation with Coefficient of Variation. This ensures that the
bias and variability ratios are not cross-correlated,
which otherwise may occur when e.g. the precipitation inputs are biased.

The Modified Kling-Gupta Efficiency (\eqn{KGE'}) is estimated as follows:
\deqn{
KGE' = 1 - \sqrt{(r - 1)^2 + (\beta - 1)^2 + (\gamma - 1)^2}
}
where:
\itemize{
\item \eqn{r} is the linear Pearson correlation coefficient between
observed and simulated values
\item \eqn{\beta = \mu_{sim} / \mu_{obs}} is the ratio of the
means (bias ratio)
\item \eqn{
        \gamma = \frac{\sigma_{sim} / \mu_{sim}}{\sigma_{sim} / \mu_{sim}}
        } is the ratio of the Coefficients of Variation (variability ratio)
}
}
\note{
Unlike the Nash–Sutcliffe Efficiency (\link{nse}), the KGE does not have an
inherent benchmark such as "mean flow", and \eqn{KGE' = 0} does not
correspond to a baseline performance.
Therefore, \eqn{KGE'} values should not be interpreted as "good" or "bad"
based solely on their sign or magnitude.
Instead, users are encouraged to examine the individual components
(\eqn{r}, \eqn{\beta}, \eqn{\gamma})
to understand the nature of model performance and consider defining
explicit benchmarks based on the study context.

For further discussion, see Knoben et al. (2019), who caution against
directly translating NSE-based interpretation thresholds to KGE.
}
\examples{
library(tidyhydro)
data(avacha)

# Supply truth and predictions as bare column names
kge2012(avacha, obs, sim)

# Or as numeric vectors
kge2012_vec(avacha$obs, avacha$sim)
}
\references{
Kling, H., Fuchs, M., & Paulin, M. (2012). Runoff conditions in the upper
Danube basin under an ensemble of climate change scenarios.
Journal of Hydrology, 424–425, 264–277.
\doi{10.1016/j.jhydrol.2012.01.011}

Knoben, W. J. M., Freer, J. E., & Woods, R. A. (2019).
Technical note: Inherent benchmark or not? Comparing Nash–Sutcliffe and
Kling–Gupta efficiency scores. Hydrology and Earth System Sciences, 23,
4323–4331. \doi{10.5194/hess-23-4323-2019}
}
\seealso{
Other numeric metrics: 
\code{\link{kge}()},
\code{\link{mse}()},
\code{\link{nse}()},
\code{\link{pbias}()},
\code{\link{press}()},
\code{\link{rmse}()},
\code{\link{sfe}()}

Other accuracy metrics: 
\code{\link{kge}()},
\code{\link{mse}()},
\code{\link{nse}()},
\code{\link{pbias}()},
\code{\link{press}()},
\code{\link{rmse}()},
\code{\link{sfe}()}
}
\concept{accuracy metrics}
\concept{numeric metrics}
\keyword{gof}
