% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kge.R
\name{kgelog_low}
\alias{kgelog_low}
\title{Log-transformed Modified Kling-Gupta Efficiency}
\usage{
kgelog_low(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
and \code{estimate} arguments.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{kgelog_low_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
Calculate the modified Kling-Gupta Efficiency (\emph{Kling et al., 2012}),
aka \eqn{KGE'}. Dimensionless (from \eqn{-\infty} to 1).
\code{kge2012()} assesses the accuracy of
simulated data by considering correlation, bias, and variability relative
to observed data.
}
\examples{
library(tidyhydro)

# Supply truth and predictions as bare column names
kgelog_low(avacha, obs, sim)

# Or as numeric vectors
kgelog_low_vec(avacha$obs, avacha$sim)
}
\seealso{
Other KGE variants: 
\code{\link{kge}()},
\code{\link{kge2012}()}
}
\concept{KGE variants}
\keyword{gof}
