% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\alias{mse.data.frame}
\alias{mse_vec}
\title{Mean Squared Error (MSE)}
\usage{
mse(data, ...)

\method{mse}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)

mse_vec(truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
and \code{estimate} arguments.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{mse_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
The MSE is a metric that evaluates the goodness of fit between model
simulations and observations (Fisher, 1920). Measured in the squared
units of \code{truth} and \code{estimate} and can vary from \eqn{-\infty} to
\eqn{+\infty}.
}
\details{
The MSE is estimated as follows (Clark et al., 2021):
\deqn{
MSE = \frac{1}{n} \sum_{i=1}^{n}{(sim_i - obs_i)^2}
}
where:
\itemize{
\item \eqn{sim} defines model simulations at time step \eqn{i}
\item \eqn{obs} defines model observations at time step \eqn{i}
}
}
\examples{
library(tidyhydro)
data(avacha)

# Supply truth and predictions as bare column names
mse(avacha, obs, sim)

# Or as numeric vectors
mse_vec(avacha$obs, avacha$sim)
}
\references{
Fisher, R. A. (1920). Accuracy of observation, a mathematical
examination of the methods of determining, by the mean error and
by the mean square error. Monthly Notices of the Royal Astronomical
Society, 80, 758–770. \url{https://doi.org/10.1093/mnras/80.8.758}

Clark, M. P., Vogel, R. M., Lamontagne, J. R., Mizukami, N.,
Knoben, W. J. M., Tang, G., Gharari, S., Freer, J. E., Whitfield,
P. H., Shook, K. R., & Papalexiou, S. M. (2021). The Abuse of Popular
Performance Metrics in Hydrologic Modeling. Water Resources Research, 57(9),
e2020WR029001. \url{https://doi.org/10.1029/2020WR029001}
}
\seealso{
Other numeric metrics: 
\code{\link{kge}()},
\code{\link{kge2012}()},
\code{\link{nse}()},
\code{\link{pbias}()},
\code{\link{press}()},
\code{\link{sfe}()}

Other accuracy metrics: 
\code{\link{kge}()},
\code{\link{kge2012}()},
\code{\link{nse}()},
\code{\link{pbias}()},
\code{\link{press}()},
\code{\link{sfe}()}
}
\concept{accuracy metrics}
\concept{numeric metrics}
