% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{nse}
\alias{nse}
\alias{nse.data.frame}
\alias{nse_vec}
\title{Nash-Sutcliffe efficiency (NSE)}
\usage{
nse(data, ...)

\method{nse}{data.frame}(data, truth, estimate, na_rm = TRUE, performance = FALSE, ...)

nse_vec(truth, estimate, na_rm = TRUE, performance = FALSE, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
and \code{estimate} arguments.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{performance}{The optional column, indicating should the \code{nse()} return
metric interpretation. See details.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{nse_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
Calculate the Nash-Sutcliffe efficiency (Nash & Sutcliffe, 1970).
Dimensionless (from \eqn{-\infty} to 1). \code{nse()} indicates how well the plot
of observed versus simulated data fits the 1:1 line.
}
\details{
The Nash-Sutcliffe efficiency is a normalized statistic that determines
the relative magnitude of the residual variance ("noise") compared to the
measured data variance ("information"; Nash and Sutcliffe, 1970). According
to Moriasi et al. (2015) the metric interpretation can be as follows:
\itemize{
\item \strong{Excellent}/\strong{Very Good} -- \code{nse()} > 0.8
\item \strong{Good} -- 0.6 <= \code{nse()} <= 0.8
\item \strong{Satisfactory} -- 0.5 < \code{nse()} < 0.6
\item \strong{Poor} -- \code{nse()} <= 0.5
}
}
\examples{
library(tidyhydro)
data(avacha)

# Supply truth and predictions as bare column names
nse(avacha, obs, sim)

# Or as numeric vectors
nse_vec(avacha$obs, avacha$sim)
}
\references{
Nash, J. E., & Sutcliffe, J. V. (1970). River flow forecasting through
conceptual models part I — A discussion of principles. Journal of Hydrology,
10(3), 282–290. \url{https://doi.org/10.1016/0022-1694(70)90255-6}

Moriasi, D. N., Gitau, M. W., Pai, N., & Daggupati, P. (2015). Hydrologic
and Water Quality Models: Performance Measures and Evaluation Criteria.
Transactions of the ASABE, 58(6), 1763–1785.
\url{https://doi.org/10.13031/trans.58.10715}
}
\seealso{
Other numeric metrics: 
\code{\link{kge}()},
\code{\link{kge2012}()},
\code{\link{mse}()},
\code{\link{pbias}()},
\code{\link{press}()},
\code{\link{sfe}()}

Other accuracy metrics: 
\code{\link{kge}()},
\code{\link{kge2012}()},
\code{\link{mse}()},
\code{\link{pbias}()},
\code{\link{press}()},
\code{\link{sfe}()}
}
\concept{accuracy metrics}
\concept{numeric metrics}
