% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbias.R
\name{pbias}
\alias{pbias}
\alias{pbias.data.frame}
\alias{pbias_vec}
\title{Percent bias (PBIAS)}
\usage{
pbias(data, ...)

\method{pbias}{data.frame}(data, truth, estimate, na_rm = TRUE, performance = FALSE, ...)

pbias_vec(truth, estimate, na_rm = TRUE, performance = FALSE, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
and \code{estimate} arguments.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{performance}{The optional column, indicating should the \code{pbias()}
return metric interpretation. See details.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{pbias_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
Percent bias (PBIAS)
}
\details{
PBIAS is the deviation of data being evaluated, expressed as a percentage.
It measures the average tendency of the simulated data to be larger or
smaller than their observed counterparts (Moriasi et al., 2015). The optimal
value of PBIAS is 0.0, with low-magnitude values indicating accurate model
simulation.
Positive values indicate model underestimation bias,
and negative values indicate model overestimation bias (Gupta et al., 1999)
According to Moriasi et al. (2015) the metric interpretation can be as
follows:
\itemize{
\item \strong{Excellent}/\strong{Very Good} -- \code{pbias()} < ±5.0
\item \strong{Good} -- ±5.0 <= \code{pbias()} < ±10.0
\item \strong{Satisfactory} -- ±10.0 <= \code{pbias()} < ±15.0
\item \strong{Poor} -- \code{pbias()} >= ±15.0
}
}
\examples{
library(tidyhydro)
data(avacha)

# Supply truth and predictions as bare column names
pbias(avacha, obs, sim)

# Or as numeric vectors
pbias_vec(avacha$obs, avacha$sim)
}
\references{
Moriasi, D. N., Gitau, M. W., Pai, N., & Daggupati, P. (2015). Hydrologic
and Water Quality Models: Performance Measures and Evaluation Criteria.
Transactions of the ASABE, 58(6), 1763–1785.
\url{https://doi.org/10.13031/trans.58.10715}

Gupta, H. V., S. Sorooshian, and P. O. Yapo. (1999).
Status of automatic calibration for hydrologic models: Comparison with
multilevel expert calibration. J. Hydrologic Eng. 4(2): 135-143
}
\seealso{
Other numeric metrics: 
\code{\link{kge}()},
\code{\link{mse}()},
\code{\link{nse}()},
\code{\link{press}()},
\code{\link{sfe}()}

Other accuracy metrics: 
\code{\link{kge}()},
\code{\link{mse}()},
\code{\link{nse}()},
\code{\link{press}()},
\code{\link{sfe}()}
}
\concept{accuracy metrics}
\concept{numeric metrics}
