% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/press.R
\name{press}
\alias{press}
\alias{press.data.frame}
\alias{press_vec}
\title{PRediction Error Sum of Squares (PRESS)}
\usage{
press(data, ...)

\method{press}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)

press_vec(truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
and \code{estimate} arguments.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{press_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
\eqn{PRESS} is a measure of the quality of a regression model using
residuals. \eqn{PRESS} is a validation-type estimator of error that uses
the deleted residuals to provide an estimate of the prediction error.
When comparing alternate regression models, selecting the model with the
lowest value of the \eqn{PRESS} statistic is a good approach because it
means that the equation produces the least error when making new predictions
(see \emph{Helsel et al., 2020}).

It is particularly valuable in assessing multiple forms of multiple
linear regressions, but it is also useful for
simply comparing different options for a single explanatory variable in
single-variable regression models.
}
\details{
The \eqn{PRESS} is only relevant for comparisons to other regression models
with the same response variable units (\emph{Rasmunsen et al., 2009}).

It estimates as follows:
\deqn{
  PRESS = \sum_{i=1}^{n}{(sim_i - obs_i)^2}
}

where:
\itemize{
\item \eqn{sim} defines model simulations at time step \eqn{i}
\item \eqn{obs} defines model observations at time step \eqn{i}
}
}
\note{
The $PRESS$ statistic is not appropriate for comparison of models having
different transformations of response variable, e.g. linear regression and
log-transformed linear regression (\emph{Helsel et al., 2020}).
}
\examples{
library(tidyhydro)
data(avacha)

# Supply truth and predictions as bare column names
press(avacha, obs, sim)

# Or as numeric vectors
press_vec(avacha$obs, avacha$sim)
}
\references{
Rasmussen, P. P., Gray, J. R., Glysson, G. D. & Ziegler, A. C.
Guidelines and procedures for computing time-series suspended-sediment
concentrations and loads from in-stream turbidity-sensor and streamflow
data. in U.S. Geological Survey Techniques and Methods book 3, chap.
C4 53 (2009) \url{https://pubs.usgs.gov/tm/tm3c4/}.

Helsel, D. R., Hirsch, R. M., Ryberg, K. R., Archfield, S. A. &
Gilroy, E. J. Statistical Methods in Water Resources. 484 (2020)
\doi{10.3133/tm4A3}.
}
\seealso{
Other numeric metrics: 
\code{\link{kge}()},
\code{\link{kge2012}()},
\code{\link{mse}()},
\code{\link{nse}()},
\code{\link{pbias}()},
\code{\link{rmse}()},
\code{\link{sfe}()}

Other accuracy metrics: 
\code{\link{kge}()},
\code{\link{kge2012}()},
\code{\link{mse}()},
\code{\link{nse}()},
\code{\link{pbias}()},
\code{\link{rmse}()},
\code{\link{sfe}()}
}
\concept{accuracy metrics}
\concept{numeric metrics}
\keyword{regression}
