% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe.R
\name{sfe}
\alias{sfe}
\alias{sfe.data.frame}
\alias{sfe_vec}
\title{Standard Factorial Error (SFE)}
\usage{
sfe(data, ...)

\method{sfe}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)

sfe_vec(truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
and \code{estimate} arguments.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{sfe_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
Prediction standard factorial error estimated
using standard regression methods (see \emph{Herschy, 1978}).
}
\details{
The metric is widely used for assessing Sediment Rating Curves
(e.g., Hicks et al. 2020). The model is usually considered 'unacceptable'
if the \eqn{SFE > 2}, see Hicks et al. (2011). It is estimated as follows:
\deqn{SFE = \exp\left(\sqrt{\frac{1}{n} \sum_{i=1}^{n}
\left( \log\left(\frac{obs_i}{sim_i} \right) \right)^2 }\right)}
where:
\itemize{
\item \eqn{sim} defines model simulations at time step \eqn{i}
\item \eqn{obs} defines model observations at time step \eqn{i}
}
}
\examples{
library(tidyhydro)
data(avacha)

# Supply truth and predictions as bare column names
sfe(avacha, obs, sim)

# Or as numeric vectors
sfe_vec(avacha$obs, avacha$sim)
}
\references{
Herschy, R.W. 1978: Accuracy. Chapter 10 In: Herschy, R.W. (ed.)
Hydrometry - principles and practices. John Wiley and Sons, Chichester,
511 p.

Hicks, D. M., Shankar, U., McKerchar, A. I., Basher, L., Lynn, I.,
Page, M., & Jessen, M. (2011). Suspended Sediment Yields from New Zealand
Rivers. Journal of Hydrology (New Zealand), 50(1), 81–142.
\url{https://doi.org/10.3316/informit.315190637227597}

Hicks, M., Doyle, M., Watson, J., Holwerda, N., Lynch, B., Wyatt, J.,
Jones, H., & Hill, R. (2020). Measurement of Fluvial Suspended Sediment
Load and its Composition (No. 1.0.0; National Environmental Monitoring
Standards, p. 138).
\url{https://www.nems.org.nz/documents/suspended-sediment}
}
\seealso{
Other numeric metrics: 
\code{\link{kge}()},
\code{\link{kge2012}()},
\code{\link{mse}()},
\code{\link{nse}()},
\code{\link{pbias}()},
\code{\link{press}()}

Other accuracy metrics: 
\code{\link{kge}()},
\code{\link{kge2012}()},
\code{\link{mse}()},
\code{\link{nse}()},
\code{\link{pbias}()},
\code{\link{press}()}
}
\concept{accuracy metrics}
\concept{numeric metrics}
