% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyhydro-package.R
\docType{package}
\name{tidyhydro-package}
\alias{tidyhydro-package}
\title{tidyhydro: A tidy approach to hydrological analysis}
\description{
A tidy approach to hydrological analysis.
}
\examples{
\dontrun{
print("TODO")
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonmoulds/tidyhydro}
}

}
\author{
\strong{Maintainer}: Simon Moulds \email{simon.moulds@ed.ac.uk} (\href{https://orcid.org/0000-0002-7297-482X}{ORCID})

}
