% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avacha}
\alias{avacha}
\title{Mean Daily Water Discharge At Avacha River (Elizovo City)}
\source{
\itemize{
\item \url{https://gmvo.skniivh.ru/}
\item \url{https://ewds.climate.copernicus.eu/datasets/cems-glofas-historical}
}
}
\value{
\item{avacha}{a data frame}
}
\description{
Mean Daily Water Discharge At Avacha River (Elizovo City)
}
\details{
These data contain the measured (\code{obs}) mean daily water discharge
values (in \eqn{m^3/s}) at the Avacha River -- Elizovo City state gauging
station for the 2022 calendar year. They are accompanied by the GloFAS v4.0
reanalysis water discharge values for the last 24 hours (\code{sim}), derived
from
\url{https://ewds.climate.copernicus.eu/datasets/cems-glofas-historical}.

Read more about GloFAS Water Discharge reanalysis --
\url{https://confluence.ecmwf.int/display/CEMS/GloFAS+v4.0}
}
\examples{
data(avacha)
str(avacha)
}
\keyword{datasets}
