% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability.R
\name{cv}
\alias{cv}
\alias{cv.data.frame}
\alias{cv_vec}
\title{Coefficient of Variation (Cv)}
\usage{
cv(data, ...)

\method{cv}{data.frame}(data, truth, na_rm = TRUE, ...)

cv_vec(truth, na_rm = TRUE, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
and \code{estimate} arguments.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{numeric} vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{cv_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
Coefficient of Variation (Cv)
}
\examples{
library(tidyhydro)

cv

# Supply truth as bare column names
cv(avacha, obs)

# Or as numeric vectors
cv_vec(avacha$obs)
}
\concept{descriptive statistics}
\keyword{summary_stats}
