% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyhydro-package.R
\docType{package}
\name{tidyhydro-package}
\alias{tidyhydro}
\alias{tidyhydro-package}
\title{tidyhydro: Tidy Metrics for Assessing Hydrological Models Performance}
\description{
Provides tidy tools for comparing simulated and observed hydrological time series. Includes compatibility with the 'yardstick' package for model performance evaluation using commonly used metrics such as the Nash–Sutcliffe Efficiency (NSE), Kling–Gupta Efficiency (KGE), percent bias (pBIAS) and etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atsyplenkov/tidyhydro}
  \item \url{https://atsyplenkov.github.io/tidyhydro/}
  \item Report bugs at \url{https://github.com/atsyplenkov/tidyhydro/issues}
}

}
\author{
\strong{Maintainer}: Anatoly Tsyplenkov \email{atsyplenkov@fastmail.com} (\href{https://orcid.org/0000-0003-4144-8402}{ORCID}) [copyright holder]

}
\keyword{internal}
