% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyhydro-package.R
\docType{package}
\name{tidyhydro-package}
\alias{tidyhydro}
\alias{tidyhydro-package}
\title{tidyhydro: Tidy Metrics For Hydrological Modelling Quality Assessment}
\description{
Tidy tools for comparison of simulated and observed hydrological time series. yardstick compatible goodness-of-fit criteria such as NSE, KGE, pBIAS, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atsyplenkov/tidyhydro}
  \item Report bugs at \url{https://github.com/atsyplenkov/tidyhydro/issues}
}

}
\author{
\strong{Maintainer}: Anatoly Tsyplenkov \email{atsyplenkov@fastmail.com} (\href{https://orcid.org/0000-0003-4144-8402}{ORCID})

}
\keyword{internal}
